Extract menu logic from CampaignGUI into MekHQMenuBar
Currently, menu-related logic accounts for approximately 48% of the CampaignGUI class.
Extracting this code is the first step toward the future goal of creating a clearer
separation between the UI and control layers, as well as addressing existing issues
with localization.

This PR extracs all menu-related logic into a dedicated MekHQMenuBar class.
To minimize the risk of regressions given the large footprint of this change, every
method has been moved exactly as it is. The only modifications made were to rewire
access to the MekHQ, Campaign, and CampaignGUI objects using dedicated getters.
